src/requirements.txt
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
streamlit>=1.12.0
flask>=3.0.0
flask-cors>=4.0.0
